+++ NEWS.md
@@ -5,6 +5,9 @@
 - `read_src_dir()` added a `.read_domains` argument to allow users to only load
   in specific domains. If not specified, default is to load all in.
 
+- `read_src_dir()` added a `.subject_col` argument to allow users to specify the
+  name of the unique subject identifier column in the source data.
+
 ## Bug fixes
 
 - `nm_write()` now allows for special characters in the file name.